requirements.txt
asttokens==2.0.8
attrs==22.1.0
Automat==20.2.0
backcall==0.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
constantly==15.1.0
crochet==2.0.0
cryptography==38.0.1
cssselect==1.1.0
debugpy==1.6.3
decorator==5.1.1
dnspython==2.2.1
entrypoints==0.4
executing==1.1.0
filelock==3.8.0
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.0
greenlet==1.1.3
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
ipykernel==6.16.0
ipython==8.5.0
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
jupyter-core==4.11.1
jupyter_client==7.3.5
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==21.3
parsel==1.6.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
Protego==0.2.1
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
PyDispatcher==2.0.6
Pygments==2.13.0
PyJWT==2.5.0
pymongo==4.2.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pyzmq==24.0.1
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.6.3
scrapyrt==0.13.0
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.4.41
stack-data==0.5.1
tldextract==3.4.0
tornado==6.2
traitlets==5.4.0
Twisted==22.8.0
typing_extensions==4.3.0
urllib3==1.26.12
user-agent==0.1.10
w3lib==2.0.1
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zope.interface==5.4.0
